refactor : 팀스페이스 관련 컴포넌트 오류 수정
